// Projects_Assessments/Monorepo/todo-app-monorepo/packages/shared/generated-types.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type Todo = {
  readonly __typename?: 'Todo';
  readonly id: Scalars['ID']['output'];
  readonly isCompleted: Scalars['Boolean']['output'];
  readonly title: Scalars['String']['output'];
};